+++ castle/cms/static/plone.js
@@ -49,7 +49,8 @@ require([
   'castle-url/patterns/focuspoint',
   'castle-url/patterns/masonry',
   'castle-url/patterns/existing',
-  'castle-url/patterns/survey'
+  'castle-url/patterns/survey',
+  'castle-url/patterns/modallink',
 ], function($, registry, Base, Modal) {
   'use strict';